Added function get_proposition_by_criteria()
Fetches propositions from the database according to set criteria. Criteria can be set in almost any way but
with one limitation: all columns in the propositions table has to be returned (this is to keep dependent code
as simple as possible). The criteria are set using a dict where one specifies how to filter. There is no need
to set any criteria. If none are set, all propositions will be returned.